backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.0
aiofiles==24.1.0
requests==2.31.0
pydantic==2.4.2
python-multipart==0.0.6
selenium==4.15.2
webdriver-manager==4.0.1
google-generativeai==0.3.2
